+++ hc/accounts/tests/test_signup.py
@@ -2,12 +2,13 @@ from django.contrib.auth.models import User
 from django.core import mail
 from django.test import TestCase
 from django.test.utils import override_settings
-from hc.accounts.models import Project
+from hc.accounts.models import Profile, Project
 from hc.api.models import Channel, Check
 from django.conf import settings
 
 
 class SignupTestCase(TestCase):
+    @override_settings(USE_PAYMENTS=False)
     def test_it_sends_link(self):
         form = {"identity": "alice@example.org"}
 
@@ -17,6 +18,11 @@ class SignupTestCase(TestCase):
 
         # An user should have been created
         user = User.objects.get()
+
+        # A profile should have been created
+        profile = Profile.objects.get()
+        self.assertEqual(profile.sms_limit, 500)
+        self.assertEqual(profile.call_limit, 500)
 
         # And email sent
         self.assertEqual(len(mail.outbox), 1)
@@ -36,6 +42,17 @@ class SignupTestCase(TestCase):
         # A channel should have been created
         channel = Channel.objects.get()
         self.assertEqual(channel.project, project)
+
+    @override_settings(USE_PAYMENTS=True)
+    def test_it_sets_high_limits(self):
+        form = {"identity": "alice@example.org"}
+
+        self.client.post("/accounts/signup/", form)
+
+        # A profile should have been created
+        profile = Profile.objects.get()
+        self.assertEqual(profile.sms_limit, 5)
+        self.assertEqual(profile.call_limit, 0)
 
     @override_settings(REGISTRATION_OPEN=False)
     def test_it_obeys_registration_open(self):
@@ -64,3 +81,11 @@ class SignupTestCase(TestCase):
         form = {"identity": "alice at example org"}
         r = self.client.post("/accounts/signup/", form)
         self.assertContains(r, "Enter a valid email address")
+
+    def test_it_checks_length(self):
+        aaa = "a" * 300
+        form = {"identity": f"alice+{aaa}@example.org"}
+        r = self.client.post("/accounts/signup/", form)
+        self.assertContains(r, "Address is too long.")
+
+        self.assertFalse(User.objects.exists())