+++ static/js/details.js
@@ -70,7 +70,7 @@ $(function () {
             headers: {"X-CSRFToken": token},
             data: {"state": isOn ? "on" : "off"}
         });
-    })
+    });
 
     var statusUrl = document.getElementById("edit-timeout").dataset.statusUrl;
     var lastStatusText = "";
@@ -175,6 +175,13 @@ $(function () {
     // the target project:
     $("#transfer-modal").on("change", "#target-project", function() {
         $("#transfer-confirm").prop("disabled", !this.value);
+    });
+
+
+    // Enable/disable fields in the "Filtering Rules" modal
+    $("input[type=radio][name=filter_by_subject]").on("change", function() {
+        var enableInputs = this.value == "yes";
+        $(".filter-by-subject").prop("disabled", !enableInputs);
     });
 
 });